+++ verik-plugin/src/main/kotlin/io/verik/plugin/domain/VivadoTargetDomainObjectImpl.kt
@@ -14,4 +14,7 @@ class VivadoTargetDomainObjectImpl : VivadoTargetDomainObject {
     override var name: String = "vivado"
     override var part: String = ""
     override var ipConfigFiles: List<Path> = listOf()
+    override var simTop: String = ""
+    override var synthTop: String = ""
+    override var constraintsFile: Path? = null
 }